Break out tests into files based on method
